pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pulso"
version = "0.1.0"
description = "PULSO — The living emotional map of Mexico"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "httpx>=0.27",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "sse-starlette>=2.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "litellm>=1.40",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
pulso = "pulso.cli:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
